requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pdfplumber==0.10.2
langchain==0.1.12
pydantic==2.6.4
numpy==1.26.4
openai==1.14.3 
transformers==4.35.2
tensorflow==2.14.0
google-genai==2.14.0

